to avoid some warnings ...
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@9884 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/src/newread.c
@@ -1102,9 +1102,9 @@ int post_search(struct _select_def* conf, struct fileheader* fh, void* extraarg)
 {
     static char query[STRLEN];
     char ans[STRLEN], pmt[STRLEN];
-    int flag=(int)extraarg;
-    int up = flag&0x1;
-    int inmail = flag&0x2;
+    unsigned long flag=(unsigned long)extraarg;
+    int up = (int)(flag&0x1);
+    int inmail = (int)(flag&0x2);
 
     strncpy(ans, query, STRLEN);
     snprintf(pmt, STRLEN, "%sËÑÑ°ÄÚÈÝ: ", up ? "¡ü" : "¡ý");
